requirements.txt
transformers>=4.0.0
datasets>=2.0.0
scikit-learn
peft
fastapi
uvicorn[standard]
mlflow
streamlit
requests
torch
evaluate
nltk
mlflow
nltk
mlflow
nltk
comet_ml>=3.0.0
python-dotenv>=0.19.0
matplotlib>=3.3.0
torch
autogen
langchain
ollama
langchain_community
flaml[automl]
transformers
langchain-huggingface
pydantic>=2.0
